Parser: Do not store parser.consume call returns in `_` variable
I thought that it was necessary to store the `Ok` variant in some sort
of variable. It also lead to some bugs where the parser would not
actually return error from parser.consume calls. Not storing those
doesn't emit any error and parser will show warning when the `Err`
variant isn't used. Found 2 places where that was an actual problem